% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_coef.R
\name{format_coef}
\alias{format_coef}
\title{Format coefficient table}
\usage{
format_coef(df, ...)
}
\arguments{
\item{df}{A tibble from a tidy broom call}

\item{...}{Extra calls to round.}
}
\value{
A tibble
}
\description{
\code{format_coef} takes the tibble of coefficients and p.values
from a tidy call, and roudns to 3 dp and adds stars as a column.
}
\examples{

all_data_wide \%>\% filter(Treatment != "CON") \%>\%
  group_by(Host, Treatment) \%>\%
  do(tidy(glm(Inf_rate ~ MRS2, data = ., family = binomial()))) \%>\%
  format_num()
}
