% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sdr_data.R
\name{import_sdr_data}
\alias{import_sdr_data}
\title{Import SDR data files}
\usage{
import_sdr_data(dataset_dir, sheet_number = 1, skip_rows = 12,
  Trim_time = 5, ...)
}
\arguments{
\item{dataset_dir}{A string indicating the location
of the directory of xlsx files to import.}

\item{sheet_number}{An integer. The sheet number of the
excel workbook that should be imported. Defaults to
the first}

\item{skip_rows}{An integer. The number of rows of the
excel sheet to skip on import. Defaults to 12.}

\item{Trim_time}{An integer. Initial sampling times
to skip. Defaults to 5. No Sensor errors are set to missing (NA).}

\item{...}{}
}
\value{
A nested tibble by SDR, with Date, Time, Cell, and Oxygen values
}
\description{
\code{import_sdr_data} imports a directory of
SDR output files. Files must be in the xlsx format.
SDR IDs will be extracted along with the Run ID and
Oxygen consumption data.
}
\examples{
test_data <- import_metab(dataset_dir = "./friA_18NOV/test/",
                           sheet_number = 1, skip_rows = 12,
                           Trim_time = 0)
}
