% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tricoder.R
\name{import_tricoder}
\alias{import_tricoder}
\title{Import tricoder data}
\usage{
import_tricoder(dataset_dir, data_type, ext = "txt", ...)
}
\arguments{
\item{dataset_dir}{A string indicating the location
of the directory of csv files to import.}

\item{data_type}{A string indicating the type of data to import.
Can be either "size" or "offspring" at the moment.}

\item{ext}{A string indicating the extension of files to be imported.
Defaults to txt.}
}
\value{
A single tibble.
}
\description{
\code{import_tricoder} imports a directory of
csv Tricoder files. Choose whether bodysize or
offspring + death data is to be imported and the
column headings will be set correctly.
}
\examples{
test_data <- import_tricoder(dataset_dir = "./Bodysize/",
                  data_type = "size",
                  ext = "txt")
}
