% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sdr_id.R
\name{merge_sdr_id}
\alias{merge_sdr_id}
\title{Merges SDR ID and SDR data}
\usage{
merge_sdr_id(sdr_df, id_df)
}
\arguments{
\item{sdr_df}{A tibble of metab data in nested format.}

\item{id_df}{A tibble of ids with location_ID as unique.}
}
\value{
A tibble with Date, Run, SDR, Cell Details and Animal IDs.
Blanks are now indicated by BLANK.
}
\description{
\code{merge_sdr_id} takes the nested dataframe of raw
SDR data and adds to this the IDs associated with each
cell
}
\examples{
test_data_merged <- merge_sdr_id(sdr_df = test_data, id_df = first_run)
}
