% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sdr_cells.R
\name{plot_sdr_cells}
\alias{plot_sdr_cells}
\title{Plot metabolic data and LoLinR results for a single SDR}
\usage{
plot_sdr_cells(metab_df, SDR_no)
}
\arguments{
\item{metab_df}{A Tibble of metab data in nested format.
LoLinR must have  already been applied using \code{calc_metab}}

\item{SDR_no}{An integter. Indexes a given SDR in a single run.}
}
\value{
A 24 panel plot of each of best fitting models for each
cell of single SDR
}
\description{
\code{plot_metab_cells} takes the nested dataframe of LoLinR
results and plots the best fitting model for each cell of a
single SDR. Includes IDs and location information.
}
\examples{
plot_metab_cells(metab_df = test_data_merged, SDR = 8)
}
