% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_calc_metab.R
\name{re_calc_metab}
\alias{re_calc_metab}
\title{Runs LoLinR over whole SDR arrays and runs}
\usage{
re_calc_metab(nested_df, ID, min_time, max_time, metab_method = "pc",
  alpha_value = 0.8, set_slope_zero = FALSE)
}
\arguments{
\item{nested_df}{The nested df imported using import_metab().
Time must still be in minutes (e.g. Time_min)}

\item{ID}{A string. The ID of the indvidual to edit}

\item{min_time}{An integer. Use to trim data and recalc slopes}

\item{max_time}{An integer. Use to trim data and recalc slopes}

\item{metab_method}{A string. Either "pc" or "eq".
"pc"" is the suggested method by Colin for oxygen consumption data.
It is the default.}

\item{alpha_value}{Default at 0.2.}

\item{set_slope_zero}{Use to set slope for ID to missing}
}
\value{
A nested tibble with the model, summary and slopes (per hour change) as list columns.
}
\description{
\code{calc_metab} takes the LoLinR method of
estimating the best fitting change in oxygen consumption
but applies this to a nested dataframe of many SDR plate runs.
Time is converted from min to hours automatically.
}
\examples{
test_data_merged <- calc_metab(test_data_merged)
}
