% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ids_sdr.R
\name{add_ids_sdr}
\alias{add_ids_sdr}
\title{Adds SDR ID to SDR data}
\usage{
add_ids_sdr(sdr_df, id_df)
}
\arguments{
\item{sdr_df}{A tibble of metab data in nested format.}

\item{id_df}{A tibble of ids with location_ID as unique.}
}
\value{
A tibble with Date, Run, SDR, Cell Details and Animal IDs.
Blanks are now indicated by BLANK.
}
\description{
\code{add_ids_sdr} takes the nested dataframe of raw
SDR data and adds to this the IDs associated with each
cell
}
\examples{
test_data_merged <- add_ids_sdr(sdr_df = sdr_data_1, id_df = sdr_id_1)
}
