% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_metab.R, R/re_calc_metab.R
\name{calc_metab}
\alias{calc_metab}
\alias{calc_metab}
\title{Runs LoLinR over whole SDR arrays and runs}
\usage{
calc_metab(nested_df, metab_method = "pc", alpha_value = 0.2)

calc_metab(nested_df, metab_method = "pc", alpha_value = 0.2)
}
\arguments{
\item{nested_df}{The nested df imported using import_metab().
Time must still be in minutes (e.g. Time_min)}

\item{metab_method}{A string. Either "pc" or "eq".
"pc"" is the suggested method by Colin for oxygen consumption data.
It is the default.}

\item{alpha_value}{Default at 0.2.}

\item{nested_df}{The nested df imported using import_metab().
Time must still be in minutes (e.g. Time_min)}

\item{metab_method}{A string. Either "pc" or "eq".
"pc"" is the suggested method by Colin for oxygen consumption data.
It is the default.}

\item{alpha_value}{Default at 0.2.}
}
\value{
A nested tibble with the model, summary and slopes (per hour change) as list columns.

A nested tibble with the model, summary and slopes (per hour change) as list columns.
}
\description{
\code{calc_metab} takes the LoLinR method of
estimating the best fitting change in oxygen consumption
but applies this to a nested dataframe of many SDR plate runs.
Time is converted from min to hours automatically.

\code{calc_metab} takes the LoLinR method of
estimating the best fitting change in oxygen consumption
but applies this to a nested dataframe of many SDR plate runs.
Time is converted from min to hours automatically.
}
\examples{
test_data_merged <- calc_metab(test_data_merged)
test_data_merged <- calc_metab(test_data_merged)
}
