% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sdr_blank.R
\name{plot_sdr_blanks}
\alias{plot_sdr_blanks}
\title{Plot LoLinR slopes for every SDR in a run}
\usage{
plot_sdr_blanks(metab_df, dotcex = 1, bin = NULL)
}
\arguments{
\item{metab_df}{A Tibble of metab data in nested format.
LoLinR must have  already been applied using \code{calc_metab}.}
}
\value{
A plot for each of SDR showing the slope estimates
for each SDR in a run. Blanks cells are highlighted.
}
\description{
\code{plot_metab_blanks} takes the nested dataframe of LoLinR
results and plots the slope of each cell of an entire run.
PLots are split by SDR and have Blank cells highlighted.
}
\examples{
plot_metab_blanks(test_data_merged)
}
